Split out symbol implementations
